chore(app): error handling updates (#32)
#### fix(app): log errors in app.go and use explicit error handler
This change adds a handleError function which is responsible for
displaying errors and logging them in the top-level app.go file.

#### chore(app): change updatePreview to instanceChanged
This change updates `updatePreview` to `instanceChanged`. It also
removes some error checks where we always return nil. It also adds
`instanceChanged()` to a few user inputs where we should be updating the
preview, but we don't.