fix(app): ensure autoyes mode still works
https://github.com/smtg-ai/claude-squad/pull/40 broke auto yes mode. We would not press yes on prompts when we should have been.

This PR refactors things a bit so that we do press yes to prompts. For shift+tab in claude, we only make a best efford attempt to press this for the user when making new tmux sessions (happens on resume and instance creation). Otherwise, it's up to the user to press it. They have the option to not press shift+tab and let claude squad handle autoyes mode. The daemon still works as well.